# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TavrixMap Pelias Read-Only Geocoder API",
    "version": "1.0.0",
    "description": "Swagger documentation for the Docker-based Pelias geocoder stack.\n\nThe Pelias API reads from Elasticsearch only. PostGIS is never queried by\nthe API at request time. The sync worker periodically reads selected rows\nfrom the existing PostGIS database using read-only SELECT queries and\nindexes Pelias-compatible documents into Elasticsearch.\n\nData refresh flow:\n\n1. The external OSM pipeline updates PostGIS.\n2. The worker runs each enabled source on its cron schedule.\n3. The worker opens read-only PostgreSQL sessions and executes configured SELECT queries.\n4. Rows are transformed into Pelias-compatible documents.\n5. Documents are bulk indexed into Elasticsearch using deterministic IDs.\n6. Pelias API serves user requests from Elasticsearch only.\n7. Deletes are handled by source-scoped ID diffing or by blue/green reindex.\n"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Pelias API"
    }
  ],
...
  "paths": {
...
    "/worker/sync/{source}": {
      "post": {
        "tags": [
          "Worker"
        ],
        "summary": "Start source sync",
        "description": "Protected internal operation on worker port 9090.",
        "servers": [
          {
            "url": "http://postgis-readonly-sync-worker:9090",
            "description": "Internal PostGIS read-only sync worker"
          }
        ],
        "security": [
          {
            "internalToken": []
          },
          {
            "bearerInternalToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SourcePath"
          }
        ],
        "responses": {
          "202": {
            "description": "Run accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerRun"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/worker/reindex/{source}": {
      "post": {
        "tags": [
          "Worker"
        ],
        "summary": "Start source reindex",
        "servers": [
          {
            "url": "http://postgis-readonly-sync-worker:9090",
            "description": "Internal PostGIS read-only sync worker"
          }
        ],
        "security": [
          {
            "internalToken": []
          },
          {
            "bearerInternalToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SourcePath"
          }
        ],
        "responses": {
          "202": {
            "description": "Run accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerRun"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/worker/reindex-all": {
      "post": {
        "tags": [
          "Worker"
        ],
        "summary": "Start blue/green full reindex",
        "servers": [
          {
            "url": "http://postgis-readonly-sync-worker:9090",
            "description": "Internal PostGIS read-only sync worker"
          }
        ],
        "security": [
          {
            "internalToken": []
          },
          {
            "bearerInternalToken": []
          }
        ],
        "responses": {
          "202": {
            "description": "Run accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerRun"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/worker/runs": {
      "get": {
        "tags": [
          "Worker"
        ],
        "summary": "List worker runs",
        "servers": [
          {
            "url": "http://postgis-readonly-sync-worker:9090",
            "description": "Internal PostGIS read-only sync worker"
          }
        ],
        "security": [
          {
            "internalToken": []
          },
          {
            "bearerInternalToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Run history.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/WorkerRun"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/worker/runs/{run_id}": {
      "get": {
        "tags": [
          "Worker"
        ],
        "summary": "Get worker run",
        "servers": [
          {
            "url": "http://postgis-readonly-sync-worker:9090",
            "description": "Internal PostGIS read-only sync worker"
          }
        ],
        "security": [
          {
            "internalToken": []
          },
          {
            "bearerInternalToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerRun"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Run not found."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "internalToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-token"
      },
      "bearerInternalToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
...
      "SourcePath": {
        "name": "source",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "osm_pois"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid internal token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorObject": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "example": "Missing required parameter: text"
          },
          "request_id": {
            "type": "string",
            "example": "018f645e-9d55-7f36-a9b2-6f704da13f6a"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        }
      },
...
      "WorkerRun": {
        "type": "object",
        "required": [
          "run_id",
          "source",
          "mode",
          "started_at",
          "finished_at",
          "status",
          "indexed_count",
          "deleted_count",
          "failed_count",
          "error"
        ],
        "properties": {
          "run_id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "sync",
              "reindex"
            ]
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "success",
              "failed"
            ]
          },
          "indexed_count": {
            "type": "integer"
          },
          "deleted_count": {
            "type": "integer"
          },
          "failed_count": {
            "type": "integer"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}